+++ src/main/kotlin/no/nav/familie/pdf/pdf/lokalKjøring/LokalPdfService.kt
@@ -2,7 +2,6 @@ package no.nav.familie.pdf.pdf.lokalKjøring
 
 import no.nav.familie.pdf.pdf.JsonLeser.lesSøknadJson
 import no.nav.familie.pdf.pdf.PdfService
-import no.nav.familie.pdf.pdf.PdfValidator
 import no.nav.familie.pdf.pdf.domain.FeltMap
 import no.nav.familie.pdf.pdf.domain.PdfMedStandarder
 import no.nav.familie.pdf.pdf.domain.PdfStandard
