+++ src/main/kotlin/no/nav/familie/pdf/pdf/visningsvarianter/HåndterVisningsvariant.kt
@@ -1,12 +1,12 @@
 package no.nav.familie.pdf.pdf.visningsvarianter
 
 import com.itextpdf.layout.element.Div
-import no.nav.familie.pdf.pdf.PdfElementUtils.lagOverskriftH4
-import no.nav.familie.pdf.pdf.PdfElementUtils.lagTekstElement
 import no.nav.familie.pdf.pdf.PdfUtils.hentOversettelse
 import no.nav.familie.pdf.pdf.PdfUtils.håndterRekursivVerdiliste
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 import no.nav.familie.pdf.pdf.domain.VisningsVariant
+import no.nav.familie.pdf.pdf.pdfElementer.lagOverskriftH4
+import no.nav.familie.pdf.pdf.pdfElementer.lagTekstElement
 
 fun håndterVisningsvariant(
     visningsVariant: String,
